end of section 10

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistence;
 
 namespace Application.Activities.Commands
@@ -10,23 +11,40 @@ namespace Application.Activities.Commands
 
         // ve třídě budou dvě vlastnosti, resp. třídy (command a handler)
 
-        // 1) Command (MediatR.IRequest) - vstupní data pro editaci aktivity (DTO, stačí i record)
-        public class Command : IRequest // bez výstupního typu
+        // 1) Command (MediatR.IRequest) - vstupní data pro smazání aktivity (DTO, stačí i record)
+        public class Command : IRequest<Result<Unit>> // bez výstupního typu (Unit je vlastně void)
         {
             public required string Id { get; set; } // vstupní parametr - ID aktivity
         }
 
-        // 2) Handler (MediatR.IRequestHandler) - zpracování commandu
-        public class Handler(AppDbContext context) : IRequestHandler<Command>
+        // 2) Handler (MediatR.IRequestHandler) - zpracování commandu stejného typu (DeleteActivity.Command)
+        // na výstupu bude náš typ Result vracející zde void, resp. Unit
+        public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
         {
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
-                var activity = await context.Activities.FindAsync([request.Id], cancellationToken) 
-                    ?? throw new Exception("Cannot find activity");
+                // metoda FindAsync najde záznam v tabulce podle PK
+                var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
 
+                if (activity == null)
+                {
+                    // vytvoříme  a vrátíme instanci našeho objektu Result s nastavenou chybou
+                    return Result<Unit>.Failure("Activity not found", 404);
+                }
+
+                // položku označíme ke smazání
                 context.Activities.Remove(activity);
 
-                await context.SaveChangesAsync(cancellationToken);
+                // uložení všech změn do DB
+                var result = await context.SaveChangesAsync(cancellationToken) > 0; // byly ovlivněny záznamy?
+
+                if (!result)
+                {
+                    return Result<Unit>.Failure("Failed to delete the activity", 404);
+                }
+
+                // handler na mazání nevrací žádný výsledek (Unit je náhrada void)
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
